+++ docs/changelog.md
@@ -71,6 +71,11 @@
 
 ## 更新说明详情
 
+1.12.27 fix: fix stock_hk_fhpx_detail_ths interface
+
+    1. 修复 stock_hk_fhpx_detail_ths 接口
+    2. 修复 air_quality_hist 接口
+
 1.12.26 fix: fix index_stock_info interface
 
     1. 修复 index_stock_info 接口
@@ -3386,6 +3391,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.27 fix: fix stock_hk_fhpx_detail_ths interface
 
 1.12.26 fix: fix index_stock_info interface